archive/old_pipeline/nanopore_plasmid_pipeline/requirements.txt
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
jinja2>=3.0.0
biopython>=1.79
numpy>=1.21.0
plotly>=5.0.0
kaleido>=0.2.1 